+++ app/core/+mag/+app/+component/DatetimeSlider.m
@@ -28,6 +28,11 @@ classdef DatetimeSlider < matlab.ui.componentcontainer.ComponentContainer
         SelectedTime (1, 1) datetime
     end
 
+    properties (Access = private)
+        % ORIGINALTIMEZONE Time zone of original input.
+        OriginalTimeZone string {mustBeScalarOrEmpty}
+    end
+
     methods
 
         function comp = DatetimeSlider(varargin)
@@ -48,7 +53,7 @@ classdef DatetimeSlider < matlab.ui.componentcontainer.ComponentContainer
         end
 
         function time = get.SelectedTime(comp)
-            time = mag.app.internal.combineDateAndTime(comp.DatePicker.Value, comp.TimeField.Value);
+            time = mag.app.internal.combineDateAndTime(comp.DatePicker.Value, comp.TimeField.Value, TimeZone = comp.OriginalTimeZone);
         end
 
         function reset(comp)
@@ -63,6 +68,8 @@ classdef DatetimeSlider < matlab.ui.componentcontainer.ComponentContainer
         function limitsValueChanged(comp, ~, ~)
 
             comp.DatePicker.Limits = comp.Limits;
+            comp.OriginalTimeZone = comp.Limits.TimeZone;
+
             comp.updateSliderRange();
         end